Refactor user identification and add error handling

+++ src/utils/Gossip/Reaction/removePostReaction.ts
@@ -3,14 +3,30 @@
 import { PrismaClient, Prisma } from "@prisma/client";
 
 const prisma = new PrismaClient();
+const sha1 = require("sha1");
 
 export type RequestType = {
-	userId: string;
+	userEmail: string;
 	postId: string;
 };
 
-export async function removePostReaction({ userId, postId }: RequestType) {
-	const getUniqueId = userId + postId;
+export async function removePostReaction({ userEmail, postId }: RequestType) {
+	const user = await prisma.user.findUnique({
+		where: {
+			email: userEmail,
+		},
+	});
+
+	if (!user) {
+		return {
+			success: false,
+			message: "User not found",
+		};
+	}
+
+	const userId = user.id;
+
+	const getUniqueId = sha1(userId + postId);
 
 	try {
 		await prisma.postReaction.delete({
